requirements.txt
streamlit
pandas
numpy
scikit-learn
joblib
altair
pathlib
plotly.express